Started refactoring annotator to MVC and some more style fixes.
Annotator is going to follow MVC - model-view-controller structure in
the future. This means moving almost all GUI code to its own class and
at the same time making it possible to have an annotor inheritance
hierarchy on the controller side of the pattern. Model will be
represented by an array of Recordings for now.

Updated pylint to look for new style formatting in logging messages,
updated some logging messages, and did some minor style fixes.